Marker class and servlet tests work
