+++ maptivist/src/main/webapp/index.html
@@ -52,7 +52,7 @@
         <!--TODO: Connect this to the backend--> 
         <div class="form-popup-info" id="myForm">
             <span class="close" onclick="closeForm()">&times;</span>
-            <form class="form-container" id="create-form" action="/marker" method="POST" onsubmit="getId()">
+            <form class="form-container" id="create-form" action="/marker" method="POST">
                 <h1>Fill in Marker</h1>
                 <label for="email"><b>Email</b></label>
                 <br>
